requirements.txt
lxml==5.3.0
Pillow==10.4.0
opencv-python